requirements.txt
Django~=4.0.1
django-crispy-forms~=1.13.0
djangorestframework~=3.13.1
django-filter~=21.1
xlwt~=1.3.0